requirements.txt
faiss-cpu
sentence-transformers
transformers
torch
pdfplumber
python-docx
pdf2image
pytesseract
Pillow
